pyproject.toml
[project]
name = "les-mosaic-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "greenlet>=3.2.3",
    "pydantic-settings>=2.10.1",
    "python-telegram-bot>=22.2",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-aio>=1.9.0",
]
